refactor: screen styling refactoring
