+++ src/assets/components/FormSell.jsx
@@ -1,5 +1,4 @@
-import '../styles/FormSell.css'
-import iconCar from '../icons/car.svg'
+import '../styles/components/FormSell.css'
 
 function FormsSell() {
 
@@ -12,87 +11,89 @@ function FormsSell() {
             </div>
 
             <h4>
-            Seu nome completo
+                Seu nome completo
             </h4>
             <input placeholder='Insira seu nome completo' className='inputInfo'></input>
+
             <div className='alignInput'>
                 <div>
-            <h4>
-            Endereço de Email
-            </h4>
-            <input placeholder='insira.seu@email.aqui' className='inputInfo'></input>
+                    <h4>
+                        Endereço de Email
+                    </h4>
+                    <input placeholder='insira.seu@email.aqui' className='inputInfo'></input>
                 </div>
                 <div>
-            <h4>
-            Número de telefone
-            </h4>
-            <input placeholder='+(XX) XX XX XX XX' className='inputInfo'></input>
+                    <h4>
+                        Número de telefone
+                    </h4>
+                    <input placeholder='+(XX) XX XX XX XX' className='inputInfo'></input>
                 </div>
             </div>
+
             <div className='alignInput'>
                 <div>
-
-            <h4>
-            Nome do veículo
-            </h4>
-
-            <input placeholder='Insira o nome do veículo aqui' className='inputInfo'></input>
+                    <h4>
+                        Nome do veículo
+                    </h4>
+                    <input placeholder='Insira o nome do veículo aqui' className='inputInfo'></input>
                 </div>
                 <div>
-            <h4>
-            Ano de fabricação
-            </h4>
-            <input placeholder='YYYY'  className='inputInfo'></input>
+                    <h4>
+                        Ano de fabricação
+                    </h4>
+                    <input placeholder='YYYY' className='inputInfo'></input>
                 </div>
             </div>
+
             <div className='alignInput'>
                 <div>
-
-            <h4>
-           Tipo do veículo
-            </h4>
-            <select  className='dropdownInfo'>
-                <option>
-                    Selecione o tipo
-                    </option>
+                    <h4>
+                        Tipo do veículo
+                    </h4>
+                    <select className='dropdownInfo'>
+                        <option>
+                            Selecione o tipo
+                        </option>
                     </select>
                 </div>
                 <div>
 
-            <h4>
-           Marca do veículo
-            </h4>
-            <select  className='dropdownInfo'>
-                <option>
-                    Selecione a marca
-                    </option>
+                    <h4>
+                        Marca do veículo
+                    </h4>
+                    <select className='dropdownInfo'>
+                        <option>
+                            Selecione a marca
+                        </option>
                     </select>
                 </div>
+            </div>
+
+            <div className='alignInput'>
+                <div>
+                    <h4>
+                        Preço
+                    </h4>
+                    <input placeholder='R$00.00' className='inputInfo'></input>
+                </div>
+                <div>
+                    <h4>
+                        Condição
+                    </h4>
+
+                    <div className='alignRadio'>
+                        <input type='radio' />
+                        <span>Usado</span>
+                        <input type='radio' />
+                        <span>Novo</span>
                     </div>
-                    <div className='alignInput'>
-                        <div>
-
-                    <h4>
-           Preço
-            </h4>
-            <input placeholder='R$00.00'  className='inputInfo'></input>
-                        </div>
-                        <div>
-                    <h4>
-                    Condição
-            </h4>
-            <div className='alignRadio'>
-            <input type='radio'/>
-                <span>Usado</span>
-            <input type='radio'/>
-                <span>Novo</span>
-                        </div>
-                        </div>
+                </div>
             </div>
+
             <div className='submit'>
-            <button>
-                Enviar
-            </button>
+                <button>
+                    Enviar
+                </button>
             </div>
         </div>
     )
