+++ src/assets/styles/components/BuyItem.css
@@ -36,7 +36,6 @@
     gap: 1rem;
 }
 
-
 .infoMethod {
     display: flex;
     flex-direction: row;
