+++ src/assets/styles/components/DetailsItem.css
@@ -2,9 +2,6 @@
     display: flex;
     flex-direction: column;
     gap: 0.5rem;
-    border-bottom: solid;
-    border-width: 1px;
-    border-color: #d3d3d3;
     padding: 0rem 0rem 2rem 2rem;
  
 }
