+++ src/assets/styles/components/HeaderBuy.css
@@ -3,6 +3,7 @@
     flex-direction: row;
     justify-content: space-between;
 }
+
 .interactionOptions h4 {
     display: flex;
     width: 100px;
