+++ src/assets/styles/components/HeaderSell.css
