+++ src/assets/styles/components/VehicleList.css
@@ -48,7 +48,7 @@
 }
 
 .details span {
-    font-size: 0.8rem;
+    font-size: 12px;
 }
 
 .infoBatch {
