+++ src/assets/styles/pages/Buy.css
@@ -1,4 +1,4 @@
-.HeaderBuy {
+.headerBuy {
     border-bottom: solid;
     border-width: 1px;
     border-color: #d3d3d3;
