+++ src/assets/styles/pages/Sell.css
@@ -11,7 +11,4 @@
     display: flex;
     flex-direction: column;
     padding: 2rem;
-    border-bottom: solid;
-    border-width: 1px;
-    border-color: #d3d3d3;
 }
